유진

+++ src/main/java/ypjs/project/domain/Payment.java
@@ -3,6 +3,7 @@ package ypjs.project.domain;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import ypjs.project.domain.enums.PayStatus;
 
 import java.time.LocalDateTime;
 
@@ -13,7 +14,7 @@ import java.time.LocalDateTime;
 public class Payment {
     @Id @GeneratedValue
     @Column(name = "pay_id")
-    private int payId;
+    private Long payId;
 
     @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     @JoinColumn(name = "order_id")
